pyproject.toml
[tool.poetry]
name = "gemini-tweet-bot"
version = "0.1.0"
description = ""
authors = ["0oAstro <79555780+0oAstro@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
twikit = "^2.3.3"
google-genai = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
